tpm2_createpolicy: Removed lib/tpm_policy since the functions were used solely by createpolicy tool
